Added groups and launch image

+++ Shared/Counter.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class Counter: NSObject {
+class Counter {
     
     // -------------------------------------------------------------------------
     // MARK: Static declarations
@@ -24,37 +24,84 @@ class Counter: NSObject {
     
     // iOS ONLY
     #if os(iOS)
-    @discardableResult static func create(counter: Counter) -> Bool { // Register a new counter and save it to the disk
-        return FileManager.default.createFile(atPath: sharedDir.appendingPathComponent(counter.name).path, contents: "\(counter.count)\n\(Identifier(forColor: counter.color))".data(using: .utf8), attributes: nil) // Create file called as the counter with count and color as content
+    @discardableResult static func create(counter: Counter, inside dir: URL = sharedDir) -> Bool { // Register a new counter and save it to the disk
+        return FileManager.default.createFile(atPath: dir.appendingPathComponent(counter.name).path, contents: "\(counter.count)\n\(Identifier(forColor: counter.color))".data(using: .utf8), attributes: nil) // Create file called as the counter with count and color as content
+    }
+    
+    static func counters(atDirectory dir: URL) -> [Counter] {
+        
+        var parent: Counter?
+        
+        var counters_ = [Counter]()
+        do {
+            let files = try FileManager.default.contentsOfDirectory(at: dir, includingPropertiesForKeys: nil, options: .skipsHiddenFiles) // Counter files
+            
+            // Parse files and add to array
+            for file in files {
+                
+                var isDir: ObjCBool = false
+                
+                if FileManager.default.fileExists(atPath: file.path, isDirectory: &isDir) {
+                    if isDir.boolValue {
+                        // Is folder
+                        
+                        let content = counters(atDirectory: file)
+    
+                        var color = Color(from: 0)
+                        if let firstCounter = content.first { color = firstCounter.color }
+                        let counter = Counter(name: file.lastPathComponent, count: 0, color: color, childs: content)
+                        counters_.append(counter)
+                    } else {
+                        // Is counter
+                        
+                        let content = try String(contentsOf: file, encoding: .utf8).components(separatedBy: "\n")
+                        if content.count != 2 {
+                            return counters_
+                        }
+                        guard let count = Int(content[0]) else { return counters_ }
+                        guard let backColor = Int(content[1]) else { return counters_ }
+    
+                        let newCounter = Counter(name: file.lastPathComponent, count: count, color: Color(from: backColor))
+                        counters_.append(newCounter)
+                    }
+                }
+                
+            }
+    
+        } catch _ {
+            return counters_
+        }
+        
+        if dir != sharedDir && counters_.count > 0 { // If is in a group, declare parent
+            parent = Counter(name: dir.lastPathComponent, count: 0, color: counters_.last!.color, childs: counters_)
+            for newCounter in counters_ {
+                if !newCounter.isGroup {
+                    newCounter.parent = parent
+                }
+            }
+        }
+        
+        return counters_
     }
     
     static var counters: [Counter] { // Returns counters saved to the disk
         get {
-            var counters_ = [Counter]()
-            do {
-                let files = try FileManager.default.contentsOfDirectory(at: sharedDir, includingPropertiesForKeys: nil, options: .skipsHiddenFiles) // Counter files
-                // Parse files and add to array
-                for file in files {
-                    let content = try String(contentsOf: file, encoding: .utf8).components(separatedBy: "\n")
-                    if content.count != 2 {
-                        return counters_
-                    }
-                    guard let count = Int(content[0]) else { return counters_ }
-                    guard let backColor = Int(content[1]) else { return counters_ }
-                    counters_.append(Counter(name: file.lastPathComponent, count: count, color: Color(from: backColor)))
-                }
-            } catch _ {
-                return counters_
-            }
-            
-            return counters_
+            return counters(atDirectory: sharedDir)
         }
         
         set { // Remove all counters and re add
             do {
                 let files = try FileManager.default.contentsOfDirectory(at: sharedDir, includingPropertiesForKeys: nil, options: .skipsHiddenFiles)
                 for file in files {
-                    try FileManager.default.removeItem(at: file)
+                    
+                    var isDir: ObjCBool = false
+                    
+                    if FileManager.default.fileExists(atPath: file.path, isDirectory: &isDir) {
+                        if !isDir.boolValue {
+                            try FileManager.default.removeItem(at: file)
+                        }
+                    }
+                    
                 }
                 
                 for counter in newValue {
@@ -92,22 +139,58 @@ class Counter: NSObject {
     
     // Configure counter
     
+    #if os(iOS)
     init(file: URL) throws {
-        let content = try String(contentsOf: file, encoding: .utf8).components(separatedBy: "\n")
-        if content.count != 2 {
-            
+        
+        self.name_ = String()
+        self.oldName = String()
+        self.count_ = Int()
+        self.color_ = UIColor()
+    
+        let fileNotFound = StringError("File \(file.absoluteString) doesn't exit.")
+    
+        var isDir: ObjCBool = false
+        if FileManager.default.fileExists(atPath: file.path, isDirectory: &isDir) {
+            if isDir.boolValue {
+                // Is folder
+                
+                let content = Counter.counters(atDirectory: file)
+                var count = 0
+                for counter_ in content {
+                    count = count+counter_.count
+                    counter_.parent = self
+                }
+                
+                var color = Color(from: 0)
+                if let firstCounter = content.first { color = firstCounter.color }
+                self.name_ = file.lastPathComponent
+                self.count_ = count
+                self.color_ = color
+                self.childs_ = content
+                self.isGroup_ = true
+            } else {
+                // Is counter
+                
+                let content = try String(contentsOf: file, encoding: .utf8).components(separatedBy: "\n")
+                if content.count != 2 {
+                    
+                }
+                
+                let cannotReadData = StringError("Cannot read data from \(file.absoluteString).")
+                
+                guard let count = Int(content[0]) else { throw cannotReadData }
+                guard let backColor = Int(content[1]) else { throw cannotReadData }
+                
+                self.name_ = file.lastPathComponent
+                self.oldName = file.lastPathComponent
+                self.count_ = count
+                self.color_ = Color(from: backColor)
+            }
+        } else {
+            throw fileNotFound
         }
-        
-        let cannotReadData = StringError("Cannot read data from \(file.absoluteString).")
-        
-        guard let count = Int(content[0]) else { throw cannotReadData }
-        guard let backColor = Int(content[1]) else { throw cannotReadData }
-        
-        self.name_ = file.lastPathComponent
-        self.oldName = file.lastPathComponent
-        self.count_ = count
-        self.color_ = Color(from: backColor)
     }
+    #endif
     
     init(name: String, count: Int, color: UIColor) {
         self.name_ = name
@@ -115,6 +198,72 @@ class Counter: NSObject {
         self.count_ = count
         self.color_ = color
     }
+    
+    init(name: String, count: Int, color: UIColor, childs: [Counter]) {
+        self.name_ = name
+        self.oldName = name
+        self.count_ = count
+        self.color_ = color
+        #if os(iOS)
+        self.childs_ = childs
+        self.isGroup_ = true
+            
+        // There is no child without parent 😊
+        for child in childs {
+            child.parent = self
+        }
+        #endif
+    }
+    
+    #if os(iOS)
+    var groupDirectory: URL? { // Group folder if it's
+        if isGroup {
+            return Counter.sharedDir.appendingPathComponent(name)
+        } else {
+            return nil
+        }
+    }
+    
+    var parent: Counter? // Parent group
+    private var isGroup_ = false
+    var isGroup: Bool {
+        return isGroup_
+    }
+
+    // If is a folder, content of the folder
+    private var childs_ = [Counter]()
+    var childs: [Counter] {
+        get {
+            return childs_
+        }
+        
+        set { // Remove all couters and re add
+            guard let dir = groupDirectory else {
+                return
+            }
+            
+            do {
+                let files = try FileManager.default.contentsOfDirectory(at: dir, includingPropertiesForKeys: nil, options: .skipsHiddenFiles)
+                for file in files {
+                    var isDir: ObjCBool = false
+    
+                    if FileManager.default.fileExists(atPath: file.path, isDirectory: &isDir) {
+                        if !isDir.boolValue {
+                            try FileManager.default.removeItem(at: file)
+                        }
+                    }
+                }
+                
+                for counter in newValue {
+                    FileManager.default.createFile(atPath: dir.appendingPathComponent(counter.name).path, contents: "\(counter.count)\n\(Identifier(forColor: counter.color))".data(using: .utf8), attributes: nil)
+                }
+            } catch _ {
+                return
+            }
+        }
+    }
+    
+    #endif
     
     // For iOS, Background Color, and for Watch OS, Text Color
     private var color_: UIColor
@@ -147,6 +296,20 @@ class Counter: NSObject {
     private var count_: Int
     var count: Int {
         get {
+            
+            #if os(iOS)
+            if isGroup {
+                // If is group, return count of all childs
+                
+                var count_ = 0
+                for child in childs {
+                    count_ = count_+child.count
+                }
+                
+                return count_
+            }
+            #endif
+            
             return count_
         }
         
@@ -169,7 +332,11 @@ class Counter: NSObject {
     // Row in counters array
     var row: Int {
         var i = 0
-        for counter in Counter.counters {
+        var array = Counter.counters
+        if let parent = parent {
+            array = parent.childs
+        }
+        for counter in array {
             if counter.count_ == count_ && counter.name_ == name_ {
                 return i
             }
@@ -181,8 +348,27 @@ class Counter: NSObject {
     
     // Remove counter from disk
     func remove() {
-        
+    
+        if isGroup {
+            guard let dir = groupDirectory else { return }
+            try? FileManager.default.removeItem(at: dir)
+        }
+    
         var i = 0
+    
+        if let parent = parent {
+            for counter in parent.childs {
+                if counter.count == count && counter.name == name {
+                    parent.childs.remove(at: i)
+                    break
+                }
+    
+                i = i+1
+            }
+    
+            return
+        }
+    
         for counter in Counter.counters {
             if counter.count == count && counter.name == name {
                 Counter.counters.remove(at: i)
@@ -198,8 +384,24 @@ class Counter: NSObject {
     // Save changes
     @discardableResult func sync() -> Bool {
         
-        let sharedFile = Counter.sharedDir.appendingPathComponent(name)
-        let oldFile = Counter.sharedDir.appendingPathComponent(oldName)
+        var dir = Counter.sharedDir
+        
+        #if os(iOS)
+            if isGroup {
+                // Don't do anything if is a group
+                return false
+            }
+            
+            if let parent = parent {
+                if let dir_ = parent.groupDirectory {
+                    dir = dir_
+                }
+            }
+        #endif
+        
+        
+        let sharedFile = dir.appendingPathComponent(name)
+        let oldFile = dir.appendingPathComponent(oldName)
         
         do {
             try FileManager.default.removeItem(at: oldFile) // Remove old file
